Re-structure claim button and mobile responsive.
